# Conflicts:
#	src/main/resources/templates/work/work-register.html

+++ src/main/java/com/app/bideo/domain/order/OrderVO.java
@@ -17,11 +17,11 @@ public class OrderVO {
     private Long auctionId;
     private String orderType;
     private String licenseType;
-    private Integer originalPrice;
-    private Integer discountAmount;
-    private Integer feeAmount;
-    private Integer totalPrice;
-    private Integer depositAmount;
+    private Long originalPrice;
+    private Long discountAmount;
+    private Long feeAmount;
+    private Long totalPrice;
+    private Long depositAmount;
     private String depositStatus;
     private LocalDateTime balanceDueAt;
     private LocalDateTime orderedAt;
