+++ src/main/java/com/app/bideo/domain/settlement/SettlementVO.java
@@ -12,9 +12,9 @@ public class SettlementVO {
     private Long id;
     private Long paymentId;
     private Long memberId;
-    private Integer preTaxAmount;
-    private Integer totalDeduction;
-    private Integer netAmount;
+    private Long preTaxAmount;
+    private Long totalDeduction;
+    private Long netAmount;
     private Integer effectiveTaxRate;
     private String status;
     private LocalDateTime approvedAt;
